+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
+## [unreleased]
+
+### Added
+- Added an icon to the layer if the layer has an active filter.
+
 ## [2025.6.2] - 2025-19-6
 
 ### Added