+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
+## [2025.7.0] - 2025-19-6
+### Fixed
+- The "Clear map" button now clears symbols and resets the state for drawings (before drawings were cleared but state not proper reset).  
+
 ## [2025.6.2] - 2025-19-6
 
 ### Added